Use enableMultiJS and make timeline JS library a module
